+++ test/screenshot/infra/lib/cbt-api.js
@@ -30,13 +30,14 @@ const Cli = require('./cli');
 const CliColor = require('./logger').colors;
 const DiffBaseParser = require('./diff-base-parser');
 const Duration = require('./duration');
+const Logger = require('./logger');
 const getStackTrace = require('./stacktrace')('CbtApi');
 
 const MDC_CBT_USERNAME = process.env.MDC_CBT_USERNAME;
 const MDC_CBT_AUTHKEY = process.env.MDC_CBT_AUTHKEY;
 const REST_API_BASE_URL = 'https://crossbrowsertesting.com/api/v3';
 const SELENIUM_SERVER_URL = `http://${MDC_CBT_USERNAME}:${MDC_CBT_AUTHKEY}@hub.crossbrowsertesting.com:80/wd/hub`;
-const {ExitCode, SELENIUM_STALLED_TIME_MS} = require('./constants');
+const {ExitCode, SELENIUM_ZOMBIE_SESSION_DURATION_MS} = require('./constants');
 
 /** @type {?Promise<!Array<!cbt.proto.CbtDevice>>} */
 let allBrowsersPromise;
@@ -54,6 +55,12 @@ class CbtApi {
      * @private
      */
     this.diffBaseParser_ = new DiffBaseParser();
+
+    /**
+     * @type {!Logger}
+     * @private
+     */
+    this.logger_ = new Logger();
 
     this.validateEnvVars_();
   }
@@ -126,12 +133,17 @@ https://crossbrowsertesting.com/account
       return allBrowsersPromise;
     }
 
-    console.log('Fetching browsers from CBT...');
+    this.logger_.debug('Fetching browsers from CBT...');
 
     const stackTrace = getStackTrace('fetchAvailableDevices');
     allBrowsersPromise = this.sendRequest_(stackTrace, 'GET', '/selenium/browsers');
 
-    return allBrowsersPromise;
+    /** @type {!Array<!cbt.proto.CbtDevice>} */
+    const allBrowsers = await allBrowsersPromise;
+
+    this.logger_.debug(`Fetched ${allBrowsers.length} browsers from CBT!`);
+
+    return allBrowsers;
   }
 
   /**
@@ -363,12 +375,12 @@ https://crossbrowsertesting.com/account
   /**
    * @return {!Promise<void>}
    */
-  async killStalledSeleniumTests() {
+  async killZombieSeleniumTests() {
     // NOTE: This only returns Selenium tests running on the authenticated CBT user's account.
     // It does NOT return Selenium tests running under other users.
     /** @type {!CbtSeleniumListResponse} */
     const listResponse = await this.sendRequest_(
-      getStackTrace('killStalledSeleniumTests'),
+      getStackTrace('killZombieSeleniumTests'),
       'GET', '/selenium?active=true&num=100'
     );
 
@@ -376,7 +388,7 @@ https://crossbrowsertesting.com/account
 
     /** @type {!Array<!CbtSeleniumInfoResponse>} */
     const infoResponses = await Promise.all(activeSeleniumTestIds.map((seleniumTestId) => {
-      const infoStackTrace = getStackTrace('killStalledSeleniumTests');
+      const infoStackTrace = getStackTrace('killZombieSeleniumTests');
       return this.sendRequest_(infoStackTrace, 'GET', `/selenium/${seleniumTestId}`);
     }));
 
@@ -389,7 +401,7 @@ https://crossbrowsertesting.com/account
       }
 
       const commandTimestampMs = new Date(lastCommand.date_issued).getTime();
-      if (!Duration.hasElapsed(SELENIUM_STALLED_TIME_MS, commandTimestampMs)) {
+      if (!Duration.hasElapsed(SELENIUM_ZOMBIE_SESSION_DURATION_MS, commandTimestampMs)) {
         continue;
       }
 
@@ -406,16 +418,27 @@ https://crossbrowsertesting.com/account
    */
   async killSeleniumTests(seleniumTestIds, silent = false) {
     await Promise.all(seleniumTestIds.map(async (seleniumTestId) => {
-      if (!silent) {
-        console.log(`${CliColor.magenta('Killing')} stalled Selenium test ${CliColor.bold(seleniumTestId)}`);
-      }
+      const log = (colorVerb) => {
+        if (silent) {
+          return;
+        }
+        console.log(`${colorVerb} Selenium session ${CliColor.bold(seleniumTestId)}`);
+      };
+
+      log(CliColor.magenta('Killing'));
+
       const stackTrace = getStackTrace('killSeleniumTests');
-      return await this.sendRequest_(stackTrace, 'DELETE', `/selenium/${seleniumTestId}`).catch((err) => {
-        if (!silent) {
-          console.warn(`${CliColor.red('Failed')} to kill stalled Selenium test ${CliColor.bold(seleniumTestId)}:`);
+      const requestPromise = this.sendRequest_(stackTrace, 'DELETE', `/selenium/${seleniumTestId}`).then(
+        () => {
+          log(CliColor.bold.magenta('Killed'));
+        },
+        (err) => {
+          log(CliColor.red('Failed'));
           console.warn(err);
         }
-      });
+      );
+
+      return await requestPromise;
     }));
   }
 
@@ -433,7 +456,7 @@ https://crossbrowsertesting.com/account
     if (this.cli_.isOffline()) {
       console.warn(
         `${CliColor.magenta('WARNING')}:`,
-        new Error('CbtApi#sendRequest_() should not be called in --offline mode')
+        new Error('CbtApi.sendRequest_() should not be called in --offline mode')
       );
       return [];
     }